Moved appData validation tests to metadata tests
